CS: Introduce test suite for trustdb.TrustDB
Fixes #2138
